+++ R/plot_histogram.r
@@ -2,32 +2,35 @@
 #'
 #' This function creates histogram for each continuous feature.
 #' @param data input data to be plotted, in either \link{data.frame} or \link{data.table} format.
+#' @param title plot title
 #' @param \dots other arguments to be passed to \link{geom_histogram}.
-#' @keywords histogramcontinuous
+#' @keywords plot_histogram
+#' @aliases HistogramContinuous
 #' @import data.table
 #' @import ggplot2
 #' @importFrom scales comma
 #' @importFrom stats na.omit
 #' @import gridExtra
-#' @export
+#' @export plot_histogram HistogramContinuous
+#' @seealso \link{geom_histogram} \link{plot_density}
 #' @examples
 #' # load library
 #' library(data.table)
 #'
 #' # plot using iris data
-#' HistogramContinuous(iris)
+#' plot_histogram(iris)
 #'
 #' # plot using random data
 #' set.seed(1)
 #' data <- cbind(sapply(1:9, function(x) {rnorm(10000, sd = 30 * x)}))
-#' HistogramContinuous(data, breaks = seq(-400, 400, length = 10))
+#' plot_histogram(data, breaks = seq(-400, 400, length = 10))
 
-HistogramContinuous <- function(data, ...) {
+plot_histogram <- function(data, title = NULL, ...) {
   if (!is.data.table(data)) {data <- data.table(data)}
   ## Stop if no continuous features
-  if (SplitColType(data)$num_continuous == 0) stop("No Continuous Features")
+  if (split_columns(data)$num_continuous == 0) stop("No Continuous Features")
   ## Get continuous features
-  continuous <- SplitColType(data)$continuous
+  continuous <- split_columns(data)$continuous
   ## Get dimension
   n <- nrow(continuous)
   p <- ncol(continuous)
@@ -48,9 +51,14 @@ HistogramContinuous <- function(data, ...) {
                    })
     ## Print plot object
     if (pages > 1) {
-      suppressWarnings(do.call(grid.arrange, c(plot, ncol = 4, nrow = 4)))
+      suppressWarnings(do.call(grid.arrange, c(plot, ncol = 4, nrow = 4, top = title)))
     } else {
-      suppressWarnings(do.call(grid.arrange, plot))
+      suppressWarnings(do.call(grid.arrange, c(plot, top = title)))
     }
   }
+}
+
+HistogramContinuous <- function(data, title = NULL, ...) {
+  .Deprecated("plot_histogram")
+  plot_histogram(data = data, title = title, ...)
 }