+++ R/split_columns.r
@@ -1,8 +1,9 @@
-#' Split data into discrete and continuous
+#' Split data into discrete and continuous parts
 #'
 #' This function splits the input data into two \link{data.table} objects: discrete and continuous. A feature is continuous if \code{is.numeric} returns \code{TRUE}.
 #' @param data input data to be split, in either \link{data.frame} or \link{data.table} format.
-#' @keywords splitcoltype
+#' @keywords split_columns
+#' @aliases SplitColType
 #' @details Features with all missing values will be dropped from the output data, but will be counted towards the column count.
 #' @details The elements in the output list will have the same class as the input data.
 #' @return \code{discrete} all discrete features
@@ -11,16 +12,16 @@
 #' @return \code{num_continuous} number of continuous features
 #' @return \code{num_all_missing} number of features with no observations (all values are missing)
 #' @import data.table
-#' @export
+#' @export split_columns SplitColType
 #' @examples
-#' output <- SplitColType(iris)
+#' output <- split_columns(iris)
 #' output$discrete
 #' output$continuous
 #' output$num_discrete
 #' output$num_continuous
 #' output$num_all_missing
 
-SplitColType <- function(data) {
+split_columns <- function(data) {
   ## Check if input is data.table
   is_data_table <- is.data.table(data)
   ## Detect input data class
@@ -50,4 +51,9 @@ SplitColType <- function(data) {
       "num_all_missing" = n_all_missing
     )
   )
+}
+
+SplitColType <- function(data) {
+  .Deprecated("split_columns")
+  split_columns(data)
 }